modified:   src/data/stories.tsx 	new file:   src/data/stories/a-christmas-gift/christmasGift.tsx 	new file:   src/data/stories/anniversary-surprise/anniversarySurprise.tsx 	new file:   src/data/stories/index.tsx 	new file:   src/data/stories/new-beginnings/newBeginnings.tsx 	modified:   src/types/index.ts
