+++ src/data/stories/index.tsx
@@ -1,0 +1,10 @@
+import { Story } from '../../types/index';
+import { christmasGift } from './a-christmas-gift/christmasGift';
+import { newBeginnings } from './new-beginnings/newBeginnings';
+import { anniversarySurprise } from './anniversary-surprise/anniversarySurprise';
+
+export const stories: Story[] = [
+  christmasGift,
+  newBeginnings,
+  anniversarySurprise
+];
